add find_leaf trait method
